Break out casting to own module
